requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.8.0
boto3==1.36.0
botocore==1.36.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
jose==1.0.0
mysql-connector-python==9.1.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
